Initial public commit
Enough people liked my little declarative arugment parser for bash
scripts that I figured I would make a public release of it.

Unfortunately, that required me to fix typos in comments and update
my documentation (well, actually, to write a README.md)

So, since I work in AI these days, I figured I would use an AI to
help write the README.md - and it mostly worked.  I mean, it worked
far better than I expected and yet made a number of really silly
errors too.  But, I think I have enough now that it is worth actually
making that first commit.

Note that this mechanism is something I use in two ways - this is
just the simple "embedded" way that does not really require anything
to be installed.  I still need to update the docs such that I can
show the shared code way - which is a bit more involved due to
bash-isms and the need to have something installed somewhere.
The version here is fully self-contained.

Will have to see how this all works out as my larger, installed version
is really not ready for public release and needs a bit of extra care
to get there.  (It has been my internal play-thing for years/decades...
and thus has a lot of cruft in it that I would want to clean up first)