Add zsh tab completion support
zsh, the default on MacOS these days, is somewhat different for tab
completion.  It actually has some extra features too, such as
having option descriptions in the completion data.

Since we are declarative, we support this and use the first help
line as the short summary of the option help in addition to showing
the defined default value.

Also improve the documentation as to how to set up tab completion
for ArguBASH commands for the user.