Fix such that we can have values with spaces and "=" in them
The code had assumed that the default values would never have "=" in them and that any spaces would be not significant to the declaration.  However, if you want a default value to have an '=' in it, it would not work.  This fixes that such that only the first '=' in a default setting is significant and others are just part of the value.